+++ components/Dashboard/AdminDashboardPage/RoleTags/index.tsx
@@ -1,4 +1,4 @@
-import { Role } from '@helpers/global'
+import { Role } from '@helpers/types'
 
 const RoleTags = ({ role }: RoleProps) => {
   let roleClass: string
