+++ components/Dashboard/AdminDashboardPage/UsersTable/index.tsx
@@ -2,7 +2,7 @@ import { memo, useCallback, useState } from 'react'
 import dynamic from 'next/dynamic'
 import useSWR from 'swr'
 import Avatar from '@elements/Avatar'
-import { Role, User } from '@helpers/global'
+import { Role, User } from '@helpers/types'
 import { useUser } from '@helpers/user'
 import RoleTags from '../RoleTags'
 import SearchUser from './SearchUser'
