+++ components/Utils/Footer/ConstitutionModal/index.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react'
 import Modal from '@elements/Modal'
-import { ModalProps } from '@helpers/global'
+import { ModalProps } from '@helpers/types'
 
 const ConstitutionModal = (props: ModalProps) => (
   <Modal
