+++ components/Utils/ImageGrid/index.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import { memo, useEffect, useState } from 'react'
-import { ImageProps } from '@helpers/global'
+import { ImageProps } from '@helpers/types'
 
 const randomise = () => {
   const options = [-1, 0, 1]
